Run the program as an interpreter
Before, it was running as a sort of compiler: it assumed every byte was
an instruction, and pre-parsed every byte then ran them in order. That
doesn't work, because CHIP-8 is an interpreter. Control flow can move
all over the memory space, and some of the bytes in memory are
variably-sized sprites, so we can't pre-parse them without running
through the actual control flow.

Now the program correctly uses the stack pointer and PC (program
counter) to run the program's commands.

As part of this, there is now a `trace` subcommand that will verbosely
execute each command that it understands, printing what it's doing each
time. It panics when it reaches a command it doesn't understand.