Display a sprite on screen
This introduces the concept of a framebuffer (hope I'm using that term
correctly), which is a simple 1-dimensional vector that stores the RGB
value for any given (x, y) coordinate on the screen.

The framebuffer is data that's changed by various methods, then handed
off to `Display` to be shown on screen.

The framebuffer covers 64x32 logical pixels, and translates those behind
the scenes to 640x320 (10x scale) physical pixels on screen. This is
because 64x32 is tiny on any modern screen.