Add a `Register` newtype
This should prevent the issue in
https://github.com/gabebw/chip8/pull/25, where I confused register
values with normal bytes.

The conversion from u8 to Register is a `From` that panics, not a
`TryFrom`, because it should *never* be higher than VF in a program, and
in fact I'm not sure it's possible with the current parsing method.
Failing to convert to a `Register` *should* crash the program, rather
than being a possible failure mode that other code must handle.