Introduce `run` method in tests
It simply runs every instruction in a given program and returns the
final `State`. The DRW tests had to change because they relied on
executing all but the final instruction (which was actually sprite
data). Now the DRW tests have fully runnable programs that jump (JP)
past the initial sprite data.